changes in the stats and each enemy can have it's own level
Now hp is measured from 0 to 1 (like 0% to 100%)

+++ Battles/BattleStatsDisplayer/BattleStatsDisplayer.cs
@@ -49,9 +49,9 @@ public class BattleStatsDisplayer : Control
 		_nameLabel.Text = Stats.Name;
 	}
 	
-	private void UpdateHpBar(int hp)
+	private void UpdateHpBar(float hp)
 	{
-		_hpBar.UpdateBar((float) hp / (float) Stats.MaxHp);
+		_hpBar.UpdateBar(hp / Stats.MaxHp);
 	}
 	
 	private void UpdateXpBar(int xp)
